+++ frostmourne-monitor/src/main/java/com/autohome/frostmourne/monitor/dao/mybatis/frostmourne/repository/IDataNameRepository.java
@@ -21,5 +21,7 @@ public interface IDataNameRepository {
 
     Optional<DataName> findByName(String dataName);
 
+    List<DataName> findByNames(List<String> dataNames);
+
     List<DataName> find(String datasourceType, Long datasourceId);
 }