+++ frostmourne-vue/src/router/routes/alarm.js
@@ -22,6 +22,18 @@ const alarmRoutes = [
     name: 'alert-log',
     component: () => import('@/views/alarm/alert-log.vue'),
     meta: { title: '我的消息' }
+  },
+  {
+    path: 'alert-template.view',
+    name: 'alert-template',
+    component: () => import('@/views/alarm/alert-template.vue'),
+    meta: { title: '消息模板' }
+  },
+  {
+    path: 'service-info.view',
+    name: 'service-info',
+    component: () => import('@/views/alarm/service-info.vue'),
+    meta: { title: '服务信息' }
   }
 ]
 export default alarmRoutes