+++ frostmourne-vue/src/views/alarm/edit.vue
@@ -10,12 +10,30 @@
               </el-form-item>
             </el-col>
             <el-col :span="12">
-              <el-form-item label>
+              <el-form-item>
                 <el-switch v-model="form.status" active-value="OPEN" active-text="开启" inactive-value="CLOSE" inactive-text="关闭" />
               </el-form-item>
             </el-col>
           </el-row>
-
+          <el-row>
+            <el-col :span="12">
+              <el-form-item label="所属服务:">
+                <el-select v-model="form.serviceInfo.id" reserve-keyword placeholder="请选择服务">
+                  <el-option v-for="item in serviceOptions" :key="item.id" :label="item.serviceName" :value="item.id" />
+                </el-select>
+              </el-form-item>
+            </el-col>
+            <el-col :span="12">
+              <el-form-item label="风险等级:">
+                <el-select v-model="form.risk_level" size="small" style="width:100px" placeholder="风险等级">
+                  <el-option label="提示" value="info" />
+                  <el-option label="重要" value="important" />
+                  <el-option label="紧急" value="emergency" />
+                  <el-option label="我崩了" value="crash" />
+                </el-select>
+              </el-form-item>
+            </el-col>
+          </el-row>
           <el-row>
             <el-col :span="12">
               <el-form-item label="所属对象:">
@@ -44,12 +62,12 @@
       <el-tabs>
         <el-tab-pane label="数据配置">
           <el-row>
-            <el-col :span="8">
+            <el-col :span="6">
               <el-form-item label="数据:" prop="metricContract.data_name">
                 <el-cascader v-model="dataValue" width="100" size="medium" :show-all-levels="false" :options="dataOptions" @change="dataChange" />
               </el-form-item>
             </el-col>
-            <el-col :span="8">
+            <el-col :span="6">
               <el-form-item v-if="dataSourceType === 'elasticsearch'" label="聚合类型:">
                 <el-select v-model="form.metricContract.aggregation_type">
                   <el-option label="count" value="count" />
@@ -57,10 +75,18 @@
                   <el-option label="min" value="min" />
                   <el-option label="max" value="max" />
                   <el-option label="sum" value="sum" />
+                  <el-option label="unique count" value="cardinality" />
+                  <el-option label="standard deviation" value="standard_deviation" />
+                  <el-option label="percentiles" value="percentiles" />
                 </el-select>
               </el-form-item>
             </el-col>
-            <el-col :span="8">
+            <el-col :span="6">
+              <el-form-item v-if="dataSourceType === 'elasticsearch' && form.metricContract.aggregation_type === 'percentiles'" label="百分比:">
+                <el-input v-model="form.metricContract.properties.percent" placeholder="例如: 90" />
+              </el-form-item>
+            </el-col>
+            <el-col :span="6">
               <el-form-item v-if="dataSourceType === 'elasticsearch' && form.metricContract.aggregation_type !== 'count'" label="聚合字段:">
                 <el-input v-model="form.metricContract.aggregation_field" />
               </el-form-item>
@@ -154,6 +180,13 @@
 
         </el-tab-pane>
         <el-tab-pane label="消息模板">
+          <el-form-item>
+            <el-select v-model.number="alertTemplateId" filterable placeholder="请选择" @change="changeAlertTemplateOptions()">
+              <el-option v-for="item in alertTemplateOptions"
+                         :key="item.id" :label="item.templateName" :value="item.id" />
+            </el-select>
+            <el-button type="primary" :disabled="alertTemplateOption==null" @click="importAlertTemplate">导入模板</el-button>
+          </el-form-item>
           <el-form-item label="消息模板:" prop="ruleContract.alert_template">
             <el-input v-model="form.ruleContract.alert_template" type="textarea" rows="5" />
           </el-form-item>
@@ -247,6 +280,8 @@ import alarmApi from '@/api/alarm.js'
 import adminApi from '@/api/admin.js'
 import { teams, search } from '@/api/user'
 import dataApi from '@/api/data.js'
+import alerttemplateApi from '@/api/alert-template.js'
+import serviceinfoApi from '@/api/service-info.js'
 
 import VueJsonPretty from 'vue-json-pretty'
 
@@ -256,6 +291,7 @@ export default {
   },
   data () {
     return {
+      referer: null,
       intervalCron: '',
       dayCron: '',
       dayCronOptions: [],
@@ -302,6 +338,9 @@ export default {
           ways: [],
           recipients: [],
           silence: 60
+        },
+        serviceInfo: {
+          id: 0
         }
       },
       rules: {
@@ -340,8 +379,18 @@ export default {
       dataValue: [],
       dataOptions: [],
       teamList: [],
-      enableSaveAnother: true
+      enableSaveAnother: true,
+      alertTemplateOptions: [],
+      alertTemplateOption: null,
+      alertTemplateId: null,
+      serviceOptionsLoading: false,
+      serviceOptions: []
     }
+  },
+  beforeRouteEnter (to, from, next) {
+    next(
+      vm => { vm.referer = from }
+    )
   },
   mounted () {
     this.initDayCronOptions()
@@ -368,23 +417,35 @@ export default {
       this.dataSourceType = this.$route.query.datasource_type
       this.form.metricContract.query_string = this.$route.query.query_string
       this.dataChange(this.dataValue)
+    } else {
+      this.initAlertTemplateOptions()
     }
+
+    this.loadServiceOptions()
   },
   methods: {
+    goBack () {
+      if (this.referer) {
+        this.$router.back()
+      } else {
+        this.$router.push({ path: '/alarm/list.view' })
+      }
+    },
+    success (message) {
+      this.$message({
+        type: 'success',
+        message: message,
+        duration: 500,
+        onClose: () => this.goBack()
+      })
+    },
     onSubmit () {
       this.$refs['form'].validate((validate) => {
         if (validate) {
           this.disableSave = false
           this.copyToProperties()
           adminApi.save(this.form)
-            .then(response => {
-              this.$message({
-                type: 'success',
-                message: '保存成功！',
-                duration: 500,
-                onClose: () => this.$router.push({ path: '/alarm/list.view' })
-              })
-            })
+            .then(response => this.success('保存成功！'))
             .catch(error => {
               this.$message({
                 type: 'success',
@@ -399,15 +460,9 @@ export default {
     },
     onSaveAnother () {
       this.copyToProperties()
+      this.form.alarm_name = this.form.alarm_name + '(copy)'
       adminApi.saveAnother(this.form)
-        .then(response => {
-          this.$message({
-            type: 'success',
-            message: '另存成功！',
-            duration: 500,
-            onClose: () => this.$router.push({ path: '/alarm/list.view' })
-          })
-        })
+        .then(response => this.success('另存成功！'))
         .catch(error => {
           console.log('另存失败:', error)
         })
@@ -423,6 +478,9 @@ export default {
       console.log('addHeader.headers -> ' + this.httpHeaders.length, this.httpHeaders)
     },
     copyToProperties () {
+      if (this.dataSourceType !== 'http') {
+        return
+      }
       this.form.metricContract.properties = {}
       this.httpHeaders.forEach(item => {
         if (item.key !== '' && item.value !== '') {
@@ -463,11 +521,14 @@ export default {
       })
     },
     onCancel () {
-      this.$router.push({ path: '/alarm/list.view' })
+      this.goBack()
     },
     getDetail (callback) {
       adminApi.findById(this.id)
         .then(response => {
+          if (response.result.serviceInfo == null) {
+            response.result.serviceInfo = { id: 0 }
+          }
           this.form = response.result
           this.copyToHeaders(this.form.metricContract.properties)
 
@@ -489,6 +550,7 @@ export default {
       if (value.length === 0) {
         this.form.metricContract.data_source_id = 0
         this.form.metricContract.data_name = ''
+        this.initAlertTemplateOptions()
         return
       }
       this.dataSourceType = value[0]
@@ -497,11 +559,13 @@ export default {
         this.form.metricContract.data_name = 'http'
         this.form.metricContract.metric_type = 'object'
         this.form.ruleContract.rule_type = 'expression'
+        this.initAlertTemplateOptions()
         return
       }
       this.form.metricContract.data_source_id = value[1]
       this.form.metricContract.data_name = value[2]
       this.form.metricContract.metric_type = ''
+      this.initAlertTemplateOptions()
     },
     tabClick (tab, event) {
       // console.log(tab, event)
@@ -588,6 +652,68 @@ export default {
         '指标同比${item.description}变化${item.percentage}%,超过阈值${PERCENTAGE_THRESHOLD}%, 当前值: ${CURRENT}, 对比值：${item.value};\r\n' +
         '</#list>'
       }
+    },
+    initAlertTemplateOptions () {
+      this.alertTemplateId = null
+      this.alertTemplateOption = null
+      var condition = {
+        templateTypeUnionCodes: ['COMMON'],
+        pageIndex: 1,
+        pageSize: 1000
+      }
+      var dataName = this.form.metricContract.data_name
+      if (dataName != null && dataName !== '') {
+        condition.templateTypeUnionCodes.push('DATA_NAME|' + dataName)
+      }
+      alerttemplateApi.findAlertTemplate(condition)
+        .then(response => {
+          this.alertTemplateOptions = response.result.list
+        }).catch(() => {
+          console.error('消息模板加载失败')
+        })
+    },
+    changeAlertTemplateOptions () {
+      this.alertTemplateOption = null
+      if (this.alertTemplateId) {
+        for (var i = 0; i < this.alertTemplateOptions.length; i++) {
+          if (this.alertTemplateId === this.alertTemplateOptions[i].id) {
+            this.alertTemplateOption = this.alertTemplateOptions[i]
+            this.disableAlertTemplateOptions = false
+            break
+          }
+        }
+      }
+    },
+    importAlertTemplate () {
+      if (this.alertTemplateOption) {
+        this.$confirm('是否确定使用选择的模板覆盖当前消息模板?', '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'info'
+        }).then(() => {
+          this.form.ruleContract.alert_template = this.alertTemplateOption.content
+        }).catch(() => {})
+      } else {
+        this.$alert('请选择一个消息模板', '提示').catch(() => {})
+      }
+    },
+    loadServiceOptions (query) {
+      this.serviceOptionsLoading = true
+      serviceinfoApi.findServiceInfo({
+        serviceName: query,
+        pageIndex: 1,
+        pageSize: 1000,
+        orderType: 'SERVICE_NAME'
+      })
+        .then(response => {
+          this.serviceOptions = response.result.list || []
+          this.serviceOptions.unshift({
+            id: 0,
+            serviceName: '选择服务'
+          })
+          this.serviceOptionsLoading = false
+        })
+        .catch(e => {})
     }
   }
 }